# swagger-gateway-spec.yaml
openapi: 3.0.0
paths:
...
  /analyzer-gateway/verifier/status-lists/{listId}/status/{index}:
    get:
      operationId: ExternalApiController_hundleGetStatus
      parameters:
        - name: listId
          required: true
          in: path
          description: Status list lisの識別子
          schema:
            type: string
        - name: index
          required: true
          in: path
          description: Status listの状態が設定されている位置情報
          schema:
            type: number
      responses:
        '200':
          description: Status listの状態取得が正常に終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Status listの状態を取得する
      tags: *ref_0
...
  /analyzer-gateway/manager/status-lists/{listId}/status/{index}:
    patch:
      operationId: ExternalApiController_handleUpdateStatus
      parameters:
        - name: listId
          required: true
          in: path
          description: Status list lisの識別子
          schema:
            type: string
        - name: index
          required: true
          in: path
          description: Status listの状態が設定されている位置情報
          schema:
            example: 20
            type: number
        - name: X-Correlation-ID
          in: header
          description: 処理識別子
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusListUpdateDto'
      responses:
        '200':
          description: Status listの状態更新が正常終了した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusListSuccessResponse'
        '400':
          description: リクエストが無効
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 認証は成功したが、この操作を実行する権限がない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: データが存在しない
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: データの検証に失敗した
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: サーバー内部エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags: *ref_0
...
info:
  title: VC Analyzer System GateWay API
  description: API documentation for the VC Analyzer System management plane.
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    ErrorServiceMetadata:
      type: object
      properties:
        serviceName:
          type: string
          description: サービス名
        version:
          type: string
          description: サービスバージョン
        timestamp:
          type: string
          description: 応答時間
        processingTimeMillis:
          type: number
          description: 処理時間
        correlationId:
          type: string
          description: 処理識別子
      required:
        - serviceName
        - version
        - timestamp
        - processingTimeMillis
        - correlationId
    ErrorMainResponse:
      type: object
      properties:
        code:
          type: string
          description: エラーコード
        message:
          type: string
          description: エラーメッセージ
        details:
          description: エラー詳細
          type: array
          items:
            type: object
      required:
        - code
        - message
        - details
    ErrorResponse:
      type: object
      properties:
        serviceMetadata:
          description: サービスメタ情報
          allOf:
            - $ref: '#/components/schemas/ErrorServiceMetadata'
        error:
          description: エラー情報
          allOf:
            - $ref: '#/components/schemas/ErrorMainResponse'
      required:
        - serviceMetadata
        - error
    StatusListResponse:
      type: object
      properties:
        listId:
          type: string
          description: 作成、更新または確認対象のリスト識別子
        index:
          type: number
          description: 確認対象のインデックス
        status:
          type: string
          description: 指定インデックスの現在のステータス文字列
        bitValue:
          type: number
          description: 指定インデックスの現在のビット値
        newBitValue:
          type: number
          description: 更新されたインデックスのビット値
        newStatus:
          type: string
          description: 更新後の新しいステータス文字列
        message:
          type: string
          description: 成功メッセージ
      required:
        - listId
    ServiceMetadataDto:
      type: object
      properties:
        serviceName:
          type: string
          description: サービス名
        version:
          type: string
          description: サービスバージョン
        timestamp:
          type: string
          description: 応答時間
        processingTimeMillis:
          type: number
          description: 処理時間
        verifableCredentialUrl:
          type: string
          description: IPFS URL(オプション）
        fetchedCid:
          type: string
          description: IPFS CID（オプション）
        createdCid:
          type: string
          description: IPFS CID（オプション）
      required:
        - serviceName
        - version
        - timestamp
        - processingTimeMillis
    StatusListSuccessResponse:
      type: object
      properties:
        payload:
          description: ペイロード
          allOf:
            - $ref: '#/components/schemas/StatusListResponse'
        serviceMetadata:
          description: サービスメタ情報
          allOf:
            - $ref: '#/components/schemas/ServiceMetadataDto'
      required:
        - payload
...
    StatusListUpdateDto:
      type: object
      properties:
        status:
          type: string
          example: revoked
      required:
        - status
...